feat: team aware arcadia presets
