+++ src/lib/simulations/tests/customBenchmark/benchmarkOrchestratorTestUtils.ts
@@ -4,12 +4,9 @@ import { runCustomBenchmarkOrchestrator } from 'lib/simulations/orchestrator/run
 import { TestInput } from 'lib/simulations/tests/simTestUtils'
 import DB from 'lib/state/db'
 import { BenchmarkForm } from 'lib/tabs/tabBenchmarks/useBenchmarksTabStore'
-import {
-  applyScoringMetadataPresets,
-  applySetConditionalPresets,
-} from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
 import { TsUtils } from 'lib/utils/TsUtils'
 import { expect } from 'vitest'
+import { applyScoringMetadataPresets, applySetConditionalPresets } from "lib/conditionals/evaluation/applyPresets";
 
 export async function expectBenchmarkResultsToMatch(
   basicSpd: number,
@@ -49,8 +46,8 @@ export async function expectBenchmarkResultsToMatch(
   applyScoringMetadataPresets(benchmarkForm)
 
   const orchestrator = await runCustomBenchmarkOrchestrator(benchmarkForm)
-  const benchmarkSimScore = orchestrator.benchmarkSimResult?.simScore!
-  const perfectionSimScore = orchestrator.perfectionSimResult?.simScore!
+  const benchmarkSimScore = orchestrator.benchmarkSimResult?.simScore
+  const perfectionSimScore = orchestrator.perfectionSimResult?.simScore
 
   console.log(benchmarkSimScore)
   console.log(perfectionSimScore)
